requirements.txt

---

## requirements.txt

```plaintext
openai-whisper
ffmpeg-python
